Create SDKs per-module rather than per-project
