+++ bsp/src/org/jetbrains/bsp/data/BspProjectDataService.scala
@@ -4,17 +4,15 @@ import com.intellij.openapi.externalSystem.model.DataNode
 import com.intellij.openapi.externalSystem.model.project.ProjectData
 import com.intellij.openapi.externalSystem.service.project.IdeModifiableModelsProvider
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.projectRoots.{JavaSdk, ProjectJdkTable, Sdk}
+import com.intellij.openapi.projectRoots.Sdk
 import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.roots.impl.LanguageLevelProjectExtensionImpl
 import com.intellij.openapi.vcs.roots.VcsRootDetector
 import com.intellij.openapi.vcs.{ProjectLevelVcsManager, VcsDirectoryMapping, VcsRoot}
 import com.intellij.openapi.vfs.{LocalFileSystem, VfsUtilCore}
 import com.intellij.pom.java.LanguageLevel
-import org.apache.commons.codec.digest.DigestUtils
 import org.jetbrains.plugins.scala.project.ProjectContext
-import org.jetbrains.plugins.scala.project.external.{JdkByHome, ScalaAbstractProjectDataService, SdkReference, SdkUtils}
-
+import org.jetbrains.plugins.scala.project.external.{ScalaAbstractProjectDataService, SdkReference, SdkUtils}
 import java.io.File
 import java.util
 import scala.collection.mutable
@@ -68,7 +66,7 @@ class BspProjectDataService extends ScalaAbstractProjectDataService[BspProjectDa
     val existingJdk = Option(ProjectRootManager.getInstance(project).getProjectSdk)
     val projectJdk =
       jdk
-        .flatMap(findOrCreateSdkFromBsp)
+        .flatMap(SdkUtils.findOrCreateSdk)
         .orElse(existingJdk)
         .orElse(SdkUtils.mostRecentJdk)
     projectJdk.foreach(ProjectRootManager.getInstance(project).setProjectSdk)
@@ -82,31 +80,6 @@ class BspProjectDataService extends ScalaAbstractProjectDataService[BspProjectDa
         languageLevel =>
           LanguageLevelProjectExtensionImpl.getInstanceImpl(project).setLanguageLevel(languageLevel)
       }
-    }
-  }
-
-  private def findOrCreateSdkFromBsp(sdkReference: SdkReference): Option[Sdk] = {
-    def createFromHome = {
-      Option(sdkReference).collect {
-        case JdkByHome(home) =>
-          val name = resolveName(home)
-          val newJdk = JavaSdk.getInstance.createJdk(name, home.toString, home.getName == "jre")
-          ProjectJdkTable.getInstance.addJdk(newJdk)
-          newJdk
-      }
-    }
-
-    SdkUtils.findProjectSdk(sdkReference).orElse(createFromHome)
-  }
-
-  private def resolveName(home: File) = {
-    val suffix = if (home.getName == "jre") home.getParentFile.getName else home.getName
-    val baseName = s"BSP_$suffix"
-    val names = ProjectJdkTable.getInstance.getAllJdks.map(_.getName)
-    if (names.contains(baseName)) {
-      baseName + DigestUtils.md5Hex(home.toString).take(10)
-    } else {
-      baseName
     }
   }
 }
